Enable tiling of convolution ops.
Convolution ops without padding can be tiled along the batch loops,
the "x-loops", and the output feature loop. Convolution with padding
can be tiled with along the batch loop for now. Implement this, and
check correctness.

Also, for loop to GPU lowering, the number of workgroups should only
be used as an argument to the gpu.launch operation and not within the
launch operation body. To avoid folders from accidently doing this,
get the value of number of workgroups from a fake allocation. This
allocation doesnt affect the generated code itself.

PiperOrigin-RevId: 301388350